password edited sucessfully. Password when changed first goes thorugh first validation for example if current pass match the db, and backend validation for the new pass to confirm it meets requirement and is not null, then the new password is set as the new encryption key. the old data is retrieved from the db and decrypted, and reencrypted with the new password. password and data is then saved into the db. after the pass change is sucessful, the handlebars would be let known that the change is sucessfull and the form will be hidden and success panel shown on the same page
for the details changed, it is also done. the data when submitted, the password is first retrieved from the db where username = session currently logged in, the password is used as the encryption key for the encryption of the new data. the encrypted data is then encrypted and stored intot the db. and success msg is shown